Convert server response (GCMReg) into FCMReg type
When an FcmRegistrationDescription or FcmTemplateRegistrationDescription
are created or updated via CreateOrUpdateRegistrationImplAsync, they are
converted into respective GCM counterparts before being sent to the
service. Because of that once request has been processed, the service
returns GCM type of registration which mismatches the
CreateOrUpdateRegistrationImplAsync return type (defined by the type of
registration parameter). To avoid that issue, GCM type from server's
response is converted into its FCM alternative when needed.