Add GCM to FCM conversion into GetEntityImplAsync
The conversion is added to prevent an issue where the server is probed
for registration description by its Id and the returned Gcm registration
mismatches FcmRegistration specified by the user as generic type,
causing type casting error.

GetEntityImplAsync() is used in GetRegistrationAsync(), which seems to
be broken as it commented out in samples and only tested for failure in
tests.