Add FCM support (#20)
* Merge changes from PR#14 Added FMC support

* Convert FCM registration into GCM when serializing

* Convert notifications from FCM to GCM when sending

Notification Hubs at the moment aren't aware of existence of
FCMNotificaiton type and cannon't handle it when it arrives from the
SDK. For that reason, the FCMNotifcation is converted into
GCMNotificiation prior to being sent.

* Convert server response (GCMReg) into FCMReg type

When an FcmRegistrationDescription or FcmTemplateRegistrationDescription
are created or updated via CreateOrUpdateRegistrationImplAsync, they are
converted into respective GCM counterparts before being sent to the
service. Because of that once request has been processed, the service
returns GCM type of registration which mismatches the
CreateOrUpdateRegistrationImplAsync return type (defined by the type of
registration parameter). To avoid that issue, GCM type from server's
response is converted into its FCM alternative when needed.

* Increase PackageVersion to 2.0.3

* Fix incorrect server response conversion order

GcmTempalteRegistrationDescription returned by server was mistakenly
converted into FcmRegistrationDescription as FcmTemplateRegistrationDescription
is a subclass of FcmRegistrationDescription and that part of hierarchy
was tested first and the test succeeded, thus never reaching
FcmTemplateRegistrationDescription.

* Increase PackageVersion to 2.1.0

* Add GCM to FCM conversion into GetEntityImplAsync

The conversion is added to prevent an issue where the server is probed
for registration description by its Id and the returned Gcm registration
mismatches FcmRegistration specified by the user as generic type,
causing type casting error.

GetEntityImplAsync() is used in GetRegistrationAsync(), which seems to
be broken as it commented out in samples and only tested for failure in
tests.

* Hide GCM functionality from SDK user, ver. 3.0.0

Marks all GCM-related types and methods as 'internal' in order to
incentivize a user to switch their code to FCM methods and types.

* Update Samples to use FCM functionality, ver.3.0.0

* Update tests to use FCM methods and types

* Fix incomplete GCM to FCM copy constructor